requirements.txt
tqdm
numpy==1.26.4
pillow
matplotlib
torch-fidelity==0.3.0
torchmetrics